refactor: extract VersionMigrations into dedicated migration.py module
 - Move `VersionMigrations` class out of `util.py` and into a new `migration.py` module to separate migration concerns from general utility helpers
 - Update `__init__.py` to import `VersionMigrations` from `.migration` instead of `.util`
 - Remove the now-unused `Awaitable` and `Callable` imports from `util.py`
 - Remove the now-unused `copy` import from `__init__.py`
 - `migration.py` imports only what it needs: `copy`, `DICT_KEYS`, `DEFAULTS`, and the `Awaitable`/`Callable` typing helpers

[Closes #38] - Refactor auth storage: move credentials out of device dict
