+++ custom_components/tesira_ttp/config_flow.py
@@ -2,7 +2,7 @@
 # Filename: \custom_components\tesira_ttp\config_flow.py                                                               #
 # Repository: tesira_ttp                                                                                               #
 # Created Date: Thursday, March 19th 2026, 12:56:52 AM                                                                 #
-# Last Modified: Saturday, May 2nd 2026, 11:59:18 PM                                                                   #
+# Last Modified: Sunday, May 3rd 2026, 1:16:50 AM                                                                      #
 # Original Author: Darnel Kumar                                                                                        #
 # Author Github: https://github.com/Darnel-K                                                                           #
 #                                                                                                                      #
@@ -37,7 +37,7 @@ import homeassistant.helpers.config_validation as cv
 
 from .const import DOMAIN, DICT_KEYS, DEFAULTS, CONFIG_MODES, SUPPORTED_BLOCKS, BLOCK_SCHEMA_DATA
 from .tesira_client import TesiraClient
-from .util import gen_device_id, gen_device_dict, gen_credential_dict, gen_entity_dict, _redact_device, _devices, _credentials, _entities, _device_name_map, _entity_name_map, TesiraTTPException, VersionMigrations
+from .util import gen_device_id, gen_device_dict, gen_credential_dict, gen_entity_dict, _redact_device, _devices, _credentials, _entities, _device_name_map, _entity_name_map, TesiraTTPException
 from .schemas import _device_schema, _hub, _entity_schema
 
 _LOGGER = logging.getLogger(__name__)
